requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
pip==23.3.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
setuptools==41.2.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
